April 08 2020: Reposition functions from main. the main shoudnt be as long and it should be more readable. added menu class, to work with main to keep it clean, and be able to handle diffrent elections.
Want to make a diffrent class that will contain test data.
Will add Junit tests for automation.

Added class ScanWithMessage, to be able to scan and send message at the
same time.

Adding a ballot box, switch statmanet is unnecessary.
all compile errors will be handled later by catch statments.
for this moment all input is asumed to be valid at all times.

at option 2 adding citizen,
removed the sub-option to locate details. it is not requiered by the
system.
if such want is arised than, we should build it as a diffrent option.
the func is working, and adding cit, will add exeptions later.